Admin component

+++ db/models/order.js
@@ -12,7 +12,7 @@ const Order = db.define('orders', {
       defaultValue: 'not submitted'
     },
     total: {
-      type: Sequelize.INTEGER
+      type: Sequelize.DECIMAL(20, 2)
     },
     email:{
       type:Sequelize.STRING
